lavf/movenc: Replace pkt.dts by pkt.pts to calculate trk->track_duration
In this case, the input video is of dynamic frame rate and we don't want to duplicate or drop frames, but the output video duration calculated by DTS is incorrect, I solved it by using PTS.
There are many UGC videos with dynamic frame rates, which are represented by PTS jumps. After transcoding with ffmpeg -vsync 0 or -vsync 2, the output video duration becomes longer(input: 3.05 -> output: 5.06). By reading the code of x264/encoder/encoder.c, I found that in order to predict the B frame, x264 needs to ensure that there are enough reference frames when DTS = 0, so the DTS of these reference frames will subtract the cache time. However, the cache time includes the part of PTS jumps, which results in the abnormal small DTS. In this case, the video duration calculated by DTS is abnormally large.

Signed-off-by: Mengyang Yuan <manuelyuan@tencent.com>